Modular backend project files
    - Configuration: settings now is distributed between `server.py`, `config.py`, `app/__init__.py` and `.flaskenv` files

        - The Host and port app Info was moved from `server.py` to`.flaskenv`

    - The API endpoints handlers were moved from `server.py` to `app/api.py` file

    - The User models were moved from `backend/models.py` to `app/users/models.py` file

    - Optionally you can run the app with a script `start_local.sh`

    - Logs: saved externally in `logs/backend`
