+++ backend/.flaskenv
@@ -1,0 +1,3 @@
+FLASK_RUN_HOST=localhost
+FLASK_RUN_PORT=5002
+FLASK_APP=server.py
