+++ backend/app/__init__.py
@@ -1,0 +1,38 @@
+from flask import Flask
+import os
+import logging
+from logging.handlers import RotatingFileHandler
+
+from config import Config as config_class
+from extensions import db, migrate
+
+
+# Flask app
+app = Flask(__name__)
+app.config.from_object(config_class)
+
+db.init_app(app)
+migrate.init_app(app, db)
+
+if not app.debug and not app.testing:
+    if app.config['LOG_TO_STDOUT']:
+        stream_handler = logging.StreamHandler()
+        stream_handler.setLevel(logging.INFO)
+        app.logger.addHandler(stream_handler)
+    else:
+        if not os.path.exists('logs'):
+            os.mkdir('logs')
+        file_handler = RotatingFileHandler('logs/backend.log',
+                                           maxBytes=10240, backupCount=10)
+        file_handler.setFormatter(logging.Formatter(
+            '%(asctime)s %(levelname)s: %(message)s '
+            '[in %(pathname)s:%(lineno)d]'))
+        file_handler.setLevel(logging.INFO)
+        app.logger.addHandler(file_handler)
+
+    app.logger.setLevel(logging.INFO)
+    app.logger.info('Backend startup')
+
+
+from app import api
+from app.users import *
