+++ backend/app/middlewares.py
@@ -1,7 +1,7 @@
 from flask import request, jsonify, g
 from functools import wraps
-from models import UserSession, User
-from extensions import db
+from app.models import UserSession, User
+
 
 def require_auth(f):
     @wraps(f)
@@ -9,11 +9,11 @@ def require_auth(f):
         token = request.headers.get('Authorization')
         if not token:
             return jsonify({'error': 'Authorization token is missing'}), 401
-        
+
         session = UserSession.query.filter_by(token=token, active=True).first()
         if not session:
             return jsonify({'error': 'Invalid or expired token'}), 403
-        
+
         user = User.query.filter_by(id=session.user_id).first()
         if not user:
             return jsonify({'error': 'User not found'}), 404
