+++ backend/app/models.py
@@ -1,0 +1,1 @@
+from app.users.models import *  # noqa
