+++ backend/config.py
@@ -1,0 +1,20 @@
+import os
+from dotenv import load_dotenv
+
+
+basedir = os.path.abspath(os.path.dirname(__file__))
+load_dotenv(os.path.join(basedir, '.env'))
+load_dotenv(os.path.join(basedir, '.flaskenv'))
+
+class Config:
+    LOG_TO_STDOUT = os.environ.get('LOG_TO_STDOUT')
+
+    SECRET_KEY = os.environ.get('SECRET_KEY') or 'you-will-never-guess'
+
+    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL') or \
+                              'sqlite:///urban_workflow.db'
+    SQLALCHEMY_TRACK_MODIFICATIONS = True
+
+    WTF_CSRF_ENABLED = True
+
+    ADMINS = ['your-email@example.com']
