+++ backend/start_local.sh
@@ -1,0 +1,1 @@
+python server.py
